fix: split auth.ts into server and client modules
