+++ frontend/lib/auth.client.ts
@@ -1,0 +1,6 @@
+export function getClientToken() {
+  return document.cookie
+    .split("; ")
+    .find((c) => c.startsWith("token="))
+    ?.split("=")[1];
+}
