+++ frontend/lib/auth.ts
@@ -4,10 +4,3 @@ export async function getToken() {
   const cookieStore = await cookies();
   return cookieStore.get("token")?.value;
 }
-
-export function getClientToken() {
-  return document.cookie
-    .split("; ")
-    .find((c) => c.startsWith("token="))
-    ?.split("=")[1];
-}
